Move java outlet example to new package structure
