+++ src/com/richard/walker/outlets/java/BritishOutlet.java
@@ -1,4 +1,4 @@
-package com.richard.walker;
+package com.richard.walker.outlets.java;
 
 public class BritishOutlet implements EuropeanOutlet{
