+++ src/com/richard/walker/outlets/java/NorthAmericanOutlet.java
@@ -1,4 +1,4 @@
-package com.richard.walker;
+package com.richard.walker.outlets.java;
 
 public interface NorthAmericanOutlet{
     Double provide110Volts();
